corrected user region migration

+++ database/migrations/2024_09_12_212222_create_user_regions_table.php
@@ -11,14 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('user_regions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('region_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('user_regions')) {
+            Schema::create('user_regions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('region_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
-
     /**
      * Reverse the migrations.
      */
